requirements.txt
beautifulsoup4==4.9.3
lxml==4.6.2
pyTelegramBotAPI==3.7.6
requests==2.25.1
bs4~=0.0.1
python-dotenv